Refactor io.api to allow easier overriding of parameter sizes
Refactor io.api to include an ioDefs.api which defines the size of
parameters which can be adjusted according to the capabilites of the
platform.  To select the API sizes to use, add either the linux/ or
rtos/ directory to the interface search path, as appropriate.

Also move all APIs into an interfaces/ directory to conform to best
practices.

Resolves: LE-16328
Change-Id: I6feb490fb024cc18aa321d1ea83d13206810e31e